Remove top-level Vcpkg object + move build tool packages
.sbt and .mill cause issues with wildcard imports in buildfiles
