+++ core/src/main/scala/com/indoorvivants/vcpkg/Linking.scala
@@ -1,0 +1,7 @@
+package com.indoorvivants.vcpkg
+
+sealed trait Linking
+object Linking {
+  case object Static extends Linking
+  case object Dynamic extends Linking
+}
