+++ mill-plugin/src/main/scala/com/indoorvivants/vcpkg/mill/package.scala
@@ -2,10 +2,11 @@ package com.indoorvivants.vcpkg
 
 import upickle.default._
 
-package object mill {
-  implicit val filesInfoRw: ReadWriter[Vcpkg.FilesInfo] =
+package object millplugin {
+  implicit val filesInfoRw: ReadWriter[FilesInfo] =
     implicitly[ReadWriter[MFilesInfo]]
       .bimap(MFilesInfo.fromVcpkgFilesInfo, _.toVcpkgFilesInfo)
 
-  implicit val depRW: upickle.default.ReadWriter[Vcpkg.Dependency] = upickle.default.macroRW[Vcpkg.Dependency]
+  implicit val depRW: upickle.default.ReadWriter[Dependency] =
+    upickle.default.macroRW[Dependency]
 }
